Fix path for phantomjs binary
I have an installation of current Thruk v2.0.6 and have found an error when launching/creating a new Report. It raises this error:
<pre>
 /usr/share/thruk/script/html2pdf.sh: line 28: phantomjs: command not found 
</pre>

I have logged the 4 variables arriving at the script (*html2pdf.sh*)
<pre>
/var/lib/thruk/reports/1.html
/var/lib/thruk/reports/1.dat.pdf
/var/lib/thruk/reports/1.log
phantomjs
</pre>

If **phantomjs** is not in the "current PATH" the scripts fail and raises the above reported error. Just specifying the full path, the script and the report finishes correctly.